replace money value object from shared library to different package and add multiple improvements

+++ src/application/src/commands.rs
@@ -1,14 +1,15 @@
+use rust_decimal::Decimal;
 use uuid::Uuid;
 
 #[derive(Clone)]
 pub struct CreateOrderCommand {
     customer_id: Uuid,
-    price: f64,
+    price: Decimal,
     items: Vec<CreateOrderItemDto>,
 }
 
 impl CreateOrderCommand {
-    pub fn new(customer_id: Uuid, price: f64, items: Vec<CreateOrderItemDto>) -> Self {
+    pub fn new(customer_id: Uuid, price: Decimal, items: Vec<CreateOrderItemDto>) -> Self {
         Self {
             customer_id,
             price,
@@ -20,7 +21,7 @@ impl CreateOrderCommand {
         self.customer_id
     }
 
-    pub fn price(&self) -> f64 {
+    pub fn price(&self) -> Decimal {
         self.price
     }
 
@@ -33,12 +34,12 @@ impl CreateOrderCommand {
 pub struct CreateOrderItemDto {
     product_id: Uuid,
     quantity: i32,
-    price: f64,
-    sub_total: f64,
+    price: Decimal,
+    sub_total: Decimal,
 }
 
 impl CreateOrderItemDto {
-    pub fn new(product_id: Uuid, quantity: i32, price: f64, sub_total: f64) -> Self {
+    pub fn new(product_id: Uuid, quantity: i32, price: Decimal, sub_total: Decimal) -> Self {
         Self {
             product_id,
             quantity,
@@ -55,11 +56,11 @@ impl CreateOrderItemDto {
         self.quantity
     }
 
-    pub fn price(&self) -> f64 {
+    pub fn price(&self) -> Decimal {
         self.price
     }
 
-    pub fn sub_total(&self) -> f64 {
+    pub fn sub_total(&self) -> Decimal {
         self.sub_total
     }
 }
